Use a single TPM context and avoid race conditions during tests
This implements a singleton to use a single instance of the
tss_espi::Context.

This also avoids race conditions by using a thread-safe mutex to access
the context.

The tests were modified to cleanup generated key handles between test
cases, preventing leftover handles to fill the TPM memory.

The QuoteData::fixture implementation (used only during tests) was
modified to flush the created key contexts when dropped.  It is not
possible to run the tests that use the QuoteData::fixture in parallel
because each instance would create new keys, exceeding the available TPM
memory.  Note that in the actual agent code, it is expected that only a
single instance of QuoteData exist.

Finally, the following changes were made to the tests/run.sh:

* Do not use tpm2-abrmd anymore, and use only the swtpm socket instead
* Stop the started processes at exit
* Use a single thread to run the tests. This is necessary to avoid test
  cases that generate TPM keys to run in parallel, which can fill the
  available TPM memory, causing unpredictable errors

With all these changes, the tests/run.sh can now be executed locally
without any special setting.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>